fix: optimize client and server components + remove Effect from client side to reduce bundle size
