+++ core/src/app/global-error.tsx
@@ -3,7 +3,7 @@ import "client-only";
 import "./globals.css";
 
 import { ZapButton } from "@/components/zap-ui/button";
-import { font } from "@/lib/font";
+import { geist } from "@/fonts";
 
 interface ErrorBoundaryProps {
   reset: () => void;
@@ -12,7 +12,7 @@ interface ErrorBoundaryProps {
 export default function GlobalError({ reset }: ErrorBoundaryProps) {
   return (
     <html lang="en">
-      <body className={`${font.className} antialiased`}>
+      <body className={`${geist.variable} antialiased`}>
         <div className="flex min-h-screen items-center px-4 py-12 sm:px-6 md:px-8 lg:px-12 xl:px-16">
           <div className="w-full space-y-6 text-center">
             <div className="space-y-3">
