+++ core/src/zap/components/ai/ai-settings-sheet.tsx
@@ -1,10 +1,3 @@
-"use client";
-
-import { zodResolver } from "@hookform/resolvers/zod";
-import { useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
-
-import { Form } from "@/components/ui/form";
 import {
   Sheet,
   SheetContent,
@@ -12,15 +5,8 @@ import {
   SheetHeader,
   SheetTitle,
 } from "@/components/ui/sheet";
-import { ApiKeyInput } from "@/zap/components/ai/api-key-input";
-import { ModelSelect } from "@/zap/components/ai/model-select";
-import { ProviderSelect } from "@/zap/components/ai/provider-select";
-import { SaveSettings } from "@/zap/components/ai/save-settings";
-import { DEFAULT_MODEL, ModelsByProvider } from "@/zap/data/ai";
-import { useAISettings } from "@/zap/hooks/ai/use-ai-settings";
-import { useInitAISettings } from "@/zap/hooks/ai/use-init-ai-settings";
-import { AIFormSchema, AIProviderIdSchema } from "@/zap/schemas/ai.schema";
-import type { AIFormValues } from "@/zap/types/ai.types";
+
+import { AISettingsSheetForm } from "./ai-settings-sheet-form";
 
 interface AISettingsSheetProps {
   open: boolean;
@@ -28,55 +14,8 @@ interface AISettingsSheetProps {
 }
 
 export function AISettingsSheet({ open, onOpenChange }: AISettingsSheetProps) {
-  const [initialApiKey, setInitialApiKey] = useState<string | null>(null);
-
-  const form = useForm<AIFormValues>({
-    resolver: zodResolver(AIFormSchema),
-    defaultValues: {
-      provider: AIProviderIdSchema.options[0],
-      model: ModelsByProvider[AIProviderIdSchema.options[0]][0],
-      apiKey: "",
-    },
-  });
-
-  const { isSaving, saveApiKey, testing, handleTestApiKey } =
-    useAISettings(form);
-
-  const { loading, apiKey, model: savedModel } = useInitAISettings(form, open);
-
-  const selectedProvider = form.watch("provider");
-
-  useEffect(() => {
-    if (apiKey) {
-      setInitialApiKey(apiKey);
-      form.setValue("apiKey", apiKey, { shouldValidate: true });
-    } else {
-      form.resetField("apiKey");
-    }
-  }, [apiKey, form]);
-
-  useEffect(() => {
-    if (selectedProvider && savedModel) {
-      form.setValue("model", savedModel, {
-        shouldValidate: true,
-      });
-    } else {
-      form.setValue("model", DEFAULT_MODEL[selectedProvider], {
-        shouldValidate: true,
-      });
-    }
-  }, [form, savedModel, selectedProvider]);
-
-  const handleSubmit = async (values: AIFormValues) => {
-    await saveApiKey(values);
-    onOpenChange(false);
-  };
-
-  const isSaveDisabled =
-    isSaving || testing || form.getValues("apiKey") === initialApiKey; // Key unchanged
-
   return (
-    <Sheet open={open} onOpenChange={onOpenChange}>
+    <Sheet onOpenChange={onOpenChange} open={open}>
       <SheetContent className="sm:max-w-md">
         <SheetHeader className="space-y-2">
           <SheetTitle>AI Settings</SheetTitle>
@@ -85,27 +24,7 @@ export function AISettingsSheet({ open, onOpenChange }: AISettingsSheetProps) {
           </SheetDescription>
         </SheetHeader>
 
-        <Form {...form}>
-          <form
-            onSubmit={form.handleSubmit(handleSubmit)}
-            className="space-y-6 px-4"
-          >
-            <ProviderSelect control={form.control} disabled={isSaving} />
-            <ModelSelect
-              control={form.control}
-              disabled={isSaving}
-              provider={selectedProvider}
-            />
-            <ApiKeyInput
-              control={form.control}
-              disabled={isSaving || loading}
-              loading={loading}
-              testing={testing}
-              handleTestApiKey={handleTestApiKey}
-            />
-            <SaveSettings isSaving={isSaving} isSaveDisabled={isSaveDisabled} />
-          </form>
-        </Form>
+        <AISettingsSheetForm onOpenChange={onOpenChange} open={open} />
       </SheetContent>
     </Sheet>
   );
