+++ core/src/zap/components/ai/save-settings.tsx
@@ -9,11 +9,11 @@ export function SaveSettings({ isSaving, isSaveDisabled }: SaveSettingsProps) {
   return (
     <div className="flex justify-end">
       <ZapButton
-        type="submit"
         className="w-full sm:w-auto"
         disabled={isSaveDisabled}
         loading={isSaving}
         loadingText="Saving..."
+        type="submit"
       >
         Save Settings
       </ZapButton>
