+++ core/src/zap/components/auth/register-form.tsx
@@ -1,17 +1,9 @@
 "use client";
 
 import { zodResolver } from "@hookform/resolvers/zod";
-import Link from "next/link";
 import { useForm } from "react-hook-form";
 import type { z } from "zod/v4";
 
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
 import {
   Form,
   FormControl,
@@ -22,19 +14,12 @@ import {
 } from "@/components/ui/form";
 import { Input } from "@/components/ui/input";
 import { ZapButton } from "@/components/zap-ui/button";
-import { SETTINGS } from "@/data/settings";
-import { cn } from "@/lib/utils";
-import { PolicyLinks } from "@/zap/components/auth/policy-links";
-import { SocialProviders } from "@/zap/components/auth/social-providers";
 import { useAuth } from "@/zap/hooks/auth/use-auth";
 import { RegisterFormSchema } from "@/zap/schemas/auth.schema";
 
 type RegisterFormValues = z.infer<typeof RegisterFormSchema>;
 
-export function RegisterForm({
-  className,
-  ...props
-}: React.ComponentPropsWithoutRef<"div">) {
+export function RegisterForm() {
   const { isInCooldown, cooldown, isSubmitting, handleRegisterSubmit } =
     useAuth();
 
@@ -49,131 +34,92 @@ export function RegisterForm({
   });
 
   return (
-    <div className={cn("flex flex-col gap-6", className)} {...props}>
-      <Card className="border shadow-none">
-        <CardHeader className="text-center">
-          <CardTitle className="text-xl">Create your account</CardTitle>
-          {SETTINGS.AUTH.ENABLE_SOCIAL_PROVIDER && (
-            <CardDescription>
-              Sign up with your Apple or Google account
-            </CardDescription>
+    <Form {...form}>
+      <form
+        className="grid gap-6"
+        onSubmit={form.handleSubmit(handleRegisterSubmit)}
+      >
+        <FormField
+          control={form.control}
+          name="name"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Name</FormLabel>
+              <FormControl>
+                <Input autoComplete="name" placeholder="Your name" {...field} />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
           )}
-        </CardHeader>
-        <CardContent>
-          <div className="grid gap-6">
-            {SETTINGS.AUTH.ENABLE_SOCIAL_PROVIDER && (
-              <>
-                <SocialProviders
-                  redirectURL={SETTINGS.AUTH.REDIRECT_URL_AFTER_SIGN_UP}
-                />
-                <div className="after:border-border relative text-center text-sm after:absolute after:inset-0 after:top-1/2 after:z-0 after:flex after:items-center after:border-t">
-                  <span className="bg-card text-muted-foreground relative z-10 px-2">
-                    Or continue with
-                  </span>
-                </div>
-              </>
-            )}
+        />
 
-            <Form {...form}>
-              <form
-                onSubmit={form.handleSubmit(handleRegisterSubmit)}
-                className="grid gap-6"
-              >
-                <FormField
-                  control={form.control}
-                  name="name"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Name</FormLabel>
-                      <FormControl>
-                        <Input
-                          placeholder="Your name"
-                          autoComplete="name"
-                          {...field}
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
+        <FormField
+          control={form.control}
+          name="email"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Email</FormLabel>
+              <FormControl>
+                <Input
+                  autoComplete="email"
+                  placeholder="you@example.com"
+                  type="email"
+                  {...field}
                 />
-                <FormField
-                  control={form.control}
-                  name="email"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Email</FormLabel>
-                      <FormControl>
-                        <Input
-                          placeholder="you@example.com"
-                          type="email"
-                          autoComplete="email"
-                          {...field}
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="password"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Password</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="password"
-                          placeholder="*********"
-                          autoComplete="new-password"
-                          {...field}
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="confirmPassword"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Confirm Password</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="password"
-                          placeholder="*********"
-                          autoComplete="new-password"
-                          {...field}
-                        />
-                      </FormControl>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <ZapButton
-                  type="submit"
-                  className="w-full"
-                  disabled={isInCooldown}
-                  loading={isSubmitting}
-                  loadingText="Creating account..."
-                >
-                  {!isInCooldown
-                    ? "Create account"
-                    : `Please wait ${cooldown}s`}
-                </ZapButton>
-              </form>
-            </Form>
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
 
-            <div className="text-center text-sm">
-              Already have an account?{" "}
-              <Link href="/login" className="underline underline-offset-4">
-                Log in
-              </Link>
-            </div>
-          </div>
-        </CardContent>
-      </Card>
-      <PolicyLinks />
-    </div>
+        <FormField
+          control={form.control}
+          name="password"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Password</FormLabel>
+              <FormControl>
+                <Input
+                  autoComplete="new-password"
+                  placeholder="*********"
+                  type="password"
+                  {...field}
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <FormField
+          control={form.control}
+          name="confirmPassword"
+          render={({ field }) => (
+            <FormItem>
+              <FormLabel>Confirm Password</FormLabel>
+              <FormControl>
+                <Input
+                  autoComplete="new-password"
+                  placeholder="*********"
+                  type="password"
+                  {...field}
+                />
+              </FormControl>
+              <FormMessage />
+            </FormItem>
+          )}
+        />
+
+        <ZapButton
+          className="w-full"
+          disabled={isInCooldown}
+          loading={isSubmitting}
+          loadingText="Creating account..."
+          type="submit"
+        >
+          {isInCooldown ? `Please wait ${cooldown}s` : "Create account"}
+        </ZapButton>
+      </form>
+    </Form>
   );
 }
