+++ core/src/zap/components/landing/hero/animated-text.tsx
@@ -1,5 +1,4 @@
 "use client";
-import "client-only";
 
 import { useEffect, useRef, useState } from "react";
