+++ core/src/zap/components/landing/solution/command-card.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { Check, Clipboard } from "lucide-react";
 import { useState } from "react";
 
@@ -29,11 +31,12 @@ export function CommandCard({ command, description }: CommandCardProps) {
           <code className="text-muted-foreground text-sm whitespace-nowrap">
             {command}
           </code>
+
           <Button
-            variant="ghost"
-            size="icon"
-            onClick={handleCopy}
             className="ml-4"
+            onClick={handleCopy}
+            size="icon"
+            variant="ghost"
           >
             {copied ? (
               <Check className="h-4 w-4 text-green-500" />
@@ -43,28 +46,29 @@ export function CommandCard({ command, description }: CommandCardProps) {
           </Button>
         </div>
 
-        <AnimatedSpan delay={1500} className="text-green-500">
+        <AnimatedSpan className="text-green-500" delay={1500}>
           <span>
             ✔ &apos;pnpm&apos; has been selected as the package manager.
           </span>
         </AnimatedSpan>
 
-        <AnimatedSpan delay={2000} className="text-green-500">
+        <AnimatedSpan className="text-green-500" delay={2000}>
           <span>✔ Copying core files.</span>
         </AnimatedSpan>
 
-        <AnimatedSpan delay={2500} className="text-green-500">
+        <AnimatedSpan className="text-green-500" delay={2500}>
           <span>✔ Installing dependencies.</span>
         </AnimatedSpan>
 
-        <AnimatedSpan delay={3000} className="text-blue-500">
+        <AnimatedSpan className="text-blue-500" delay={3000}>
           <span>✔ Running prettier on the project.</span>
         </AnimatedSpan>
 
-        <TypingAnimation delay={4000} className="text-muted-foreground">
+        <TypingAnimation className="text-muted-foreground" delay={4000}>
           Success! Project initialization completed.
         </TypingAnimation>
       </Terminal>
+
       <p className="text-muted-foreground text-sm">{description}</p>
     </div>
   );
