+++ core/src/zap/components/landing/solution/solution-list.tsx
@@ -7,8 +7,8 @@ export function SolutionList() {
     <ul className="grid gap-2">
       {BENEFITS.map((solution) => (
         <li
-          key={solution}
           className="flex items-center gap-2 text-sm md:text-base"
+          key={solution}
         >
           <Check className="text-primary h-4 w-4" />
           <span className="text-foreground">{solution}</span>
