+++ core/src/zap/components/sidebar/sidebar-header.tsx
@@ -7,8 +7,8 @@ export function SidebarHeader() {
       <div className="flex w-full items-center gap-1 px-4 lg:gap-2 lg:px-6">
         <SidebarTrigger className="-ml-1" />
         <Separator
-          orientation="vertical"
           className="mx-2 data-[orientation=vertical]:h-4"
+          orientation="vertical"
         />
         <h1 className="text-base font-medium">Application</h1>
       </div>
