+++ core/src/zap/components/sidebar/sidebar-main-section.tsx
@@ -39,7 +39,7 @@ export function SidebarMainSection({ items }: SidebarMainSectionProps) {
       <SidebarGroupLabel>Application</SidebarGroupLabel>
       <SidebarMenu>
         {items.map((item) => (
-          <Collapsible key={item.title} asChild defaultOpen={item.isActive}>
+          <Collapsible asChild defaultOpen={item.isActive} key={item.title}>
             <SidebarMenuItem>
               <SidebarMenuButton asChild tooltip={item.title}>
                 <Link href={item.url}>
