+++ core/src/zap/components/sidebar/sidebar-user.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { useRouter } from "@bprogress/next/app";
-import { Effect } from "effect";
 import {
   BadgeCheck,
   Bell,
@@ -13,7 +12,6 @@ import {
 import Link from "next/link";
 import { toast } from "sonner";
 
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -29,6 +27,7 @@ import {
   SidebarMenuItem,
   useSidebar,
 } from "@/components/ui/sidebar";
+import { UserInfo } from "@/zap/components/sidebar/user-info";
 import { authClient } from "@/zap/lib/auth/client";
 
 type MenuItem = {
@@ -69,37 +68,15 @@ export function SidebarUser({ user }: SidebarUserProps) {
   const fallback = getInitials(user.name);
 
   const handleSignOut = async () => {
-    await Effect.tryPromise({
-      try: () => authClient.signOut(),
-      catch: () => ({ error: true }),
-    })
-      .pipe(
-        Effect.match({
-          onSuccess: () => {
-            toast.success("Successfully signed out");
-            router.push("/login");
-          },
-          onFailure: () => {
-            toast.error("Failed to sign out");
-          },
-        }),
-      )
-      .pipe(Effect.runPromise)
-      .catch(() => toast.error("Failed to sign out"));
-  };
+    try {
+      await authClient.signOut();
 
-  const UserInfo = () => (
-    <>
-      <Avatar className="h-8 w-8 rounded-lg">
-        <AvatarImage src={user.avatar ?? ""} alt={user.name} />
-        <AvatarFallback className="rounded-lg">{fallback}</AvatarFallback>
-      </Avatar>
-      <div className="grid flex-1 text-left text-sm leading-tight">
-        <span className="truncate font-semibold">{user.name}</span>
-        <span className="truncate text-xs">{user.email}</span>
-      </div>
-    </>
-  );
+      toast.success("Successfully signed out");
+      router.push("/login");
+    } catch {
+      toast.error("Failed to sign out");
+    }
+  };
 
   const renderItems = (items: MenuItem[]) =>
     items.map(({ label, icon: Icon, href, onClick }) =>
@@ -124,23 +101,23 @@ export function SidebarUser({ user }: SidebarUserProps) {
         <DropdownMenu>
           <DropdownMenuTrigger asChild>
             <SidebarMenuButton
-              size="lg"
               className="data-[state=open]:bg-sidebar-accent data-[state=open]:text-sidebar-accent-foreground"
+              size="lg"
             >
-              <UserInfo />
+              <UserInfo fallback={fallback} user={user} />
               <ChevronsUpDown className="ml-auto size-4" />
             </SidebarMenuButton>
           </DropdownMenuTrigger>
 
           <DropdownMenuContent
+            align="end"
             className="w-[--radix-dropdown-menu-trigger-width] min-w-56 rounded-lg"
             side={isMobile ? "bottom" : "right"}
-            align="end"
             sideOffset={4}
           >
             <DropdownMenuLabel className="p-0 font-normal">
               <div className="flex items-center gap-2 px-1 py-1.5 text-left text-sm">
-                <UserInfo />
+                <UserInfo fallback={fallback} user={user} />
               </div>
             </DropdownMenuLabel>
